Refactor types out of the orchestration context file.
This commit moves a bunch of auxiliary types out of the orchestration context
into their own modules.
